fix: predefined timeframes generation issue
